# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  AssignDormitoriesRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/DormitoryRoleAssignmentItem'
        type: array
    type: object
...
  CreateDocumentRequest:
    properties:
      category:
        $ref: '#/definitions/DocumentCategory'
      expiry_date:
        type: string
      file_url:
        type: string
      issue_date:
        type: string
      note:
        type: string
      tenant_id:
        type: string
      title:
        type: string
    type: object
  CreateDormitoryRequest:
    properties:
      address:
        type: string
      name:
        type: string
    type: object
...
  Document:
    properties:
      category:
        $ref: '#/definitions/DocumentCategory'
      created_at:
        type: string
      dormitory_id:
        type: string
      expiry_date:
        type: string
      file_url:
        type: string
      id:
        type: string
      issue_date:
        type: string
      note:
        type: string
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  DocumentCategory:
    enum:
    - contract
    - id_card
    - house_registration
    - receipt
    - other
    type: string
    x-enum-varnames:
    - DocumentCategoryContract
    - DocumentCategoryIDCard
    - DocumentCategoryHouseRegistration
    - DocumentCategoryReceipt
    - DocumentCategoryOther
  DocumentDetail:
    properties:
      category:
        $ref: '#/definitions/DocumentCategory'
      created_at:
        type: string
      dormitory_id:
        type: string
      expiry_date:
        type: string
      file_url:
        type: string
      id:
        type: string
      issue_date:
        type: string
      note:
        type: string
      tenant_first_name:
        type: string
      tenant_id:
        type: string
      tenant_last_name:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  Dormitory:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  DormitoryRoleAssignment:
    properties:
      dormitory_id:
        type: string
      dormitory_name:
        type: string
      role_id:
        type: string
      role_name:
        type: string
    type: object
  DormitoryRoleAssignmentItem:
    properties:
      dormitory_id:
        type: string
      role_id:
        type: string
    type: object
...
  UpdateDocumentRequest:
    properties:
      category:
        $ref: '#/definitions/DocumentCategory'
      expiry_date:
        type: string
      file_url:
        type: string
      issue_date:
        type: string
      note:
        type: string
      tenant_id:
        type: string
      title:
        type: string
    type: object
  UpdateDormitoryRequest:
    properties:
      address:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
info:
  contact: {}
  title: API
paths:
...
  /documents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Document'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายการเอกสาร
      tags:
      - documents
    post:
      consumes:
      - application/json
      parameters:
      - description: Document payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Document'
      security:
      - ApiKeyAuth: []
      summary: สร้างเอกสาร
      tags:
      - documents
  /documents/{id}:
    delete:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบเอกสาร
      tags:
      - documents
    get:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DocumentDetail'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลเอกสารตาม ID
      tags:
      - documents
    put:
      consumes:
      - application/json
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Document payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Document'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขเอกสาร
      tags:
      - documents
  /documents/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Document file (PDF, image, or Word, max 10MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: อัปโหลดไฟล์เอกสาร
      tags:
      - documents
  /dormitories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Dormitory'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายชื่อหอพัก (สาขา)
      tags:
      - dormitories
    post:
      consumes:
      - application/json
      parameters:
      - description: Dormitory payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateDormitoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Dormitory'
      security:
      - ApiKeyAuth: []
      summary: สร้างหอพัก (สาขา)
      tags:
      - dormitories
  /dormitories/{id}:
    delete:
      parameters:
      - description: Dormitory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: ลบหอพัก (สาขา)
      tags:
      - dormitories
    get:
      parameters:
      - description: Dormitory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dormitory'
      security:
      - ApiKeyAuth: []
      summary: ดูข้อมูลหอพักตาม ID
      tags:
      - dormitories
    put:
      consumes:
      - application/json
      parameters:
      - description: Dormitory ID
        in: path
        name: id
        required: true
        type: string
      - description: Dormitory payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateDormitoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dormitory'
      security:
      - ApiKeyAuth: []
      summary: แก้ไขหอพัก (สาขา)
      tags:
      - dormitories
  /dormitories/mine:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Dormitory'
            type: array
      security:
      - ApiKeyAuth: []
      summary: รายชื่อหอพักที่ผู้ใช้ปัจจุบันเข้าถึงได้
      tags:
      - dormitories
  /dormitories/users/{userId}:
    get:
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/DormitoryRoleAssignment'
            type: array
      security:
      - ApiKeyAuth: []
      summary: ดูรายการหอพัก/บทบาทที่ผู้ใช้ถูกกำหนดสิทธิ์
      tags:
      - dormitories
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Dormitory assignment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/AssignDormitoriesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: กำหนดหอพักที่ผู้ใช้เข้าถึงได้ (แทนที่ทั้งหมด)
      tags:
      - dormitories
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"